combine Consts and API modules
