+++ src/implementations.jl
@@ -1,29 +1,9 @@
-const use_stdcall = startswith(basename(libmpi), "msmpi")
 
 # This is required in addition to __init__() so that we can call library at precompilation time.
 if Sys.isunix()
     Libdl.dlopen(libmpi, Libdl.RTLD_LAZY | Libdl.RTLD_GLOBAL)
 end
 
-macro mpicall(expr)
-    @assert expr isa Expr && expr.head == :call && expr.args[1] == :ccall
-
-    # On unix systems we call the global symbols to allow for LD_PRELOAD interception
-    # It can be emulated in Windows (via Libdl.dllist), but this is not fast.
-    if Sys.isunix() && expr.args[2].head == :tuple &&
-            (VERSION ≥ v"1.5-" || expr.args[2].args[1] ≠ :(:MPI_Get_library_version))
-        expr.args[2] = expr.args[2].args[1]
-    end
-
-    # Microsoft MPI uses stdcall calling convention
-    # this only affects 32-bit Windows
-    # unfortunately we need to use ccall to call Get_library_version
-    # so check using library name instead
-    if use_stdcall
-        insert!(expr.args, 3, :stdcall)
-    end
-    return esc(expr)
-end
 
 function Get_library_version()
     # There is no way to query at runtime what the length of the buffer should be.
